refactor(mitm-addon): split flow metadata key linter
