+++ crates/runner/mitm-addon/tests/fixtures/flow_metadata_key_linter/dynamic_star_args.expected.txt
@@ -1,0 +1,7 @@
+dynamic_star_args.py:2: use metadata_keys.VM_RUN_ID for flow.metadata access
+dynamic_star_args.py:3: use metadata_keys.FIREWALL_NAME for flow.metadata access
+dynamic_star_args.py:4: use metadata_keys.FIREWALL_ACTION for flow.metadata access
+dynamic_star_args.py:5: use metadata_keys.AUTH_CACHE_HIT for flow.metadata access
+dynamic_star_args.py:6: use metadata_keys.FIREWALL_ERROR for flow.metadata access
+dynamic_star_args.py:7: use metadata_keys.FIREWALL_PERMISSION for flow.metadata access
+dynamic_star_args.py:8: use metadata_keys.VM_NETWORK_LOG_PATH for flow.metadata access
